requirements.txt
fastapi
uvicorn
jinja2
python-multipart
PyPDF2
sentence-transformers
python-dotenv
numpy
scikit-learn
requests